Moving NetworkManager data into a separate file
